Finished modeling and texturing level 1
It might be a bit too intense for the PS1 though, and I might add a little turn in the hallway before the final area so we can cull out more geometry, or maybe an elevator. Otherwise I think the 30 fps target might be veerryyy tight. We'll have to see though. Lighting is next, and then fixing up the previous area, because that was very glitchy last time I checked